cleanup, nu-flink type conversions
