+++ engine/flink/components/dev-table/src/main/scala/pl/touk/nussknacker/engine/flink/table/FlinkTableComponentProvider.scala
@@ -5,7 +5,8 @@ import com.typesafe.scalalogging.LazyLogging
 import net.ceedubs.ficus.readers.ArbitraryTypeReader._
 import pl.touk.nussknacker.engine.api.component.{ComponentDefinition, ComponentProvider, NussknackerVersion}
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
-import pl.touk.nussknacker.engine.flink.table.aggregate.TableAggregation
+import pl.touk.nussknacker.engine.flink.table.FlinkTableComponentProvider.configIndependentComponents
+import pl.touk.nussknacker.engine.flink.table.aggregate.TableAggregationFactory
 import pl.touk.nussknacker.engine.flink.table.extractor.TableExtractor.extractTablesFromFlinkRuntime
 import pl.touk.nussknacker.engine.flink.table.extractor.SqlStatementReader
 import pl.touk.nussknacker.engine.flink.table.extractor.SqlStatementReader.SqlStatement
@@ -47,10 +48,7 @@ class FlinkTableComponentProvider extends ComponentProvider with LazyLogging {
     ) :: ComponentDefinition(
       tableComponentName,
       new TableSinkFactory(definition)
-    ) :: ComponentDefinition(
-      "aggregate",
-      new TableAggregation()
-    ) :: Nil
+    ) :: configIndependentComponents
   }
 
   private def extractTableDefinitionsFromSqlFileOrThrow(filePath: String) = {
@@ -81,6 +79,17 @@ class FlinkTableComponentProvider extends ComponentProvider with LazyLogging {
   override def isCompatible(version: NussknackerVersion): Boolean = true
 
   override def isAutoLoaded: Boolean = false
+
+}
+
+object FlinkTableComponentProvider {
+
+  val configIndependentComponents: List[ComponentDefinition] = List(
+    ComponentDefinition(
+      "aggregate",
+      new TableAggregationFactory()
+    )
+  )
 
 }
