+++ engine/flink/components/dev-table/src/main/scala/pl/touk/nussknacker/engine/flink/table/aggregate/TableAggregation.scala
@@ -1,181 +1,106 @@
 package pl.touk.nussknacker.engine.flink.table.aggregate
 
 import org.apache.flink.api.common.RuntimeExecutionMode
-import org.apache.flink.api.common.functions.FlatMapFunction
+import org.apache.flink.api.common.functions.{FlatMapFunction, RuntimeContext}
+import org.apache.flink.configuration.Configuration
 import org.apache.flink.streaming.api.datastream.DataStream
-import org.apache.flink.table.api.DataTypes
+import org.apache.flink.streaming.api.functions.ProcessFunction
 import org.apache.flink.table.api.Expressions.{$, call}
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment
 import org.apache.flink.types.Row
 import org.apache.flink.util.Collector
 import pl.touk.nussknacker.engine.api.VariableConstants.KeyVariableName
 import pl.touk.nussknacker.engine.api._
-import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.CustomNodeError
-import pl.touk.nussknacker.engine.api.context.transformation.{
-  DefinedEagerParameter,
-  NodeDependencyValue,
-  SingleInputDynamicComponent
-}
-import pl.touk.nussknacker.engine.api.context.{OutputVar, ValidationContext}
-import pl.touk.nussknacker.engine.api.definition._
-import pl.touk.nussknacker.engine.api.parameter.ParameterName
-import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypingResult}
+import pl.touk.nussknacker.engine.api.runtimecontext.{ContextIdGenerator, EngineRuntimeContext}
 import pl.touk.nussknacker.engine.flink.api.process.{
   AbstractLazyParameterInterpreterFunction,
   FlinkCustomNodeContext,
   FlinkCustomStreamTransformation
 }
-import pl.touk.nussknacker.engine.flink.table.aggregate.TableAggregation._
+import pl.touk.nussknacker.engine.flink.table.aggregate.TableAggregation.{
+  aggregateByInternalColumnName,
+  groupByInternalColumnName
+}
 import pl.touk.nussknacker.engine.flink.table.utils.NestedRowConversions.ColumnFlinkSchema
+import pl.touk.nussknacker.engine.flink.table.utils.TypeConversions.getFlinkTypeForNuTypeOrThrow
 import pl.touk.nussknacker.engine.flink.table.utils.{NestedRowConversions, RowConversions}
 
 object TableAggregation {
-
-  val groupByParamName: ParameterName            = ParameterName("groupBy")
-  val aggregateByParamName: ParameterName        = ParameterName("aggregateBy")
-  val aggregatorFunctionParamName: ParameterName = ParameterName("aggregator")
-  // TODO local: check this
-  val outputVarParamName: ParameterName = ParameterName(OutputVar.CustomNodeFieldName)
-
-  private val groupByParam: ParameterExtractor[LazyParameter[AnyRef]] with ParameterCreatorWithNoDependency =
-    ParameterDeclaration.lazyMandatory[AnyRef](groupByParamName).withCreator()
-
-  private val aggregateByParam: ParameterExtractor[LazyParameter[AnyRef]] with ParameterCreatorWithNoDependency =
-    ParameterDeclaration.lazyMandatory[AnyRef](aggregateByParamName).withCreator()
-
-  private val aggregatorFunctionParam = {
-    val aggregators = Aggregator.allAggregators.map(a => FixedExpressionValue(s"'${a.name}'", a.name))
-    ParameterDeclaration
-      .mandatory[String](aggregatorFunctionParamName)
-      .withCreator(
-        modify = _.copy(editor = Some(FixedValuesParameterEditor(FixedExpressionValue.nullFixedValue +: aggregators)))
-      )
-  }
-
   private val aggregateByInternalColumnName = "aggregateByInternalColumn"
   private val groupByInternalColumnName     = "groupByInternalColumn"
-
 }
 
-class TableAggregation
-    extends CustomStreamTransformer
-    with SingleInputDynamicComponent[FlinkCustomStreamTransformation] {
+class TableAggregation(
+    groupByLazyParam: LazyParameter[AnyRef],
+    aggregateByLazyParam: LazyParameter[AnyRef],
+    selectedAggregator: Aggregator,
+    nodeId: NodeId
+) extends FlinkCustomStreamTransformation
+    with Serializable {
 
-  override type State = Nothing
+  override def transform(
+      start: DataStream[Context],
+      context: FlinkCustomNodeContext
+  ): DataStream[ValueWithContext[AnyRef]] = {
+    val env = start.getExecutionEnvironment
+    // Setting batch mode to enable global window operations. If source is unbounded it will throw a runtime exception
+    env.setRuntimeMode(RuntimeExecutionMode.BATCH)
 
-  override def contextTransformation(context: ValidationContext, dependencies: List[NodeDependencyValue])(
-      implicit nodeId: NodeId
-  ): ContextTransformationDefinition = {
-    case TransformationStep(Nil, _) =>
-      NextParameters(
-        parameters = groupByParam
-          .createParameter() :: aggregateByParam.createParameter() :: aggregatorFunctionParam.createParameter() :: Nil,
-        errors = List.empty,
-        state = None
+    val tableEnv = StreamTableEnvironment.create(env)
+
+    val streamOfRows = start.flatMap(new LazyInterpreterFunction(groupByLazyParam, aggregateByLazyParam, context))
+
+    val groupByFlinkType     = getFlinkTypeForNuTypeOrThrow(groupByLazyParam.returnType)
+    val aggregateByFlinkType = getFlinkTypeForNuTypeOrThrow(aggregateByLazyParam.returnType)
+
+    val inputParametersTable = NestedRowConversions.buildTableFromRowStream(
+      tableEnv = tableEnv,
+      streamOfRows = streamOfRows,
+      columnSchema = List(
+        ColumnFlinkSchema(groupByInternalColumnName, groupByFlinkType),
+        ColumnFlinkSchema(aggregateByInternalColumnName, aggregateByFlinkType)
       )
-    case TransformationStep(
-          (`groupByParamName`, groupByParam) ::
-          (`aggregateByParamName`, aggregateByParam) ::
-          (`aggregatorFunctionParamName`, DefinedEagerParameter(aggregatorName: String, _)) :: Nil,
-          _
-        ) =>
-      val outName = OutputVariableNameDependency.extract(dependencies)
+    )
 
-      val selectedAggregator = Aggregator.allAggregators
-        .find(_.name == aggregatorName)
-        .getOrElse(throw new IllegalStateException("Aggregator not found. Should be invalid at parameter level."))
-
-      val aggregatorOutputType = selectedAggregator.outputType(aggregateByParam.returnType)
-
-      val aggregateByTypeErrors = selectedAggregator.inputTypeConstraint match {
-        case Some(typeConstraint) =>
-          if (!aggregateByParam.returnType.canBeSubclassOf(typeConstraint)) {
-            List(
-              // TODO: this is a different message from other aggregators - choose one and make it consistent for all
-              CustomNodeError(
-                s"""Invalid type: "${aggregateByParam.returnType.withoutValue.display}" for selected aggregator.
-                   |"${selectedAggregator.name}" aggregator requires type: "${typeConstraint.display}".
-                   |""".stripMargin,
-                Some(aggregateByParamName)
-              )
-            )
-          } else List.empty
-        case None => List.empty
-      }
-
-      FinalResults.forValidation(context, errors = aggregateByTypeErrors)(
-        _.withVariable(outName, value = aggregatorOutputType, paramName = Some(outputVarParamName)).andThen(
-          _.withVariable(
-            KeyVariableName,
-            value = groupByParam.returnType,
-            paramName = Some(ParameterName(KeyVariableName))
-          )
-        )
+    val groupedTable = inputParametersTable
+      .groupBy($(groupByInternalColumnName))
+      .select(
+        $(groupByInternalColumnName),
+        call(selectedAggregator.flinkFunctionName, $(aggregateByInternalColumnName)).as(aggregateByInternalColumnName)
       )
+
+    val groupedStream: DataStream[Row] = tableEnv.toDataStream(groupedTable)
+
+    groupedStream
+      .process(new NewContextRowMappingProcessFunction(context.convertToEngineRuntimeContext))
   }
 
-  override def implementation(
-      params: Params,
-      dependencies: List[NodeDependencyValue],
-      finalState: Option[State]
-  ): FlinkCustomStreamTransformation = {
+  private class NewContextRowMappingProcessFunction(
+      convertToEngineRuntimeContext: RuntimeContext => EngineRuntimeContext
+  ) extends ProcessFunction[Row, ValueWithContext[AnyRef]] {
+    @transient
+    private var contextIdGenerator: ContextIdGenerator = _
 
-    val groupByLazyParam     = groupByParam.extractValueUnsafe(params)
-    val aggregateByLazyParam = aggregateByParam.extractValueUnsafe(params)
-    val aggregatorVal        = aggregatorFunctionParam.extractValueUnsafe(params)
+    override def open(configuration: Configuration): Unit = {
+      contextIdGenerator = convertToEngineRuntimeContext(getRuntimeContext).contextIdGenerator(nodeId.toString)
+    }
 
-    val aggregator = Aggregator.allAggregators
-      .find(_.name == aggregatorVal)
-      .getOrElse(
-        throw new IllegalStateException("Specified aggregator not found. Should be invalid at parameter level.")
-      )
+    override def processElement(
+        value: Row,
+        ctx: ProcessFunction[Row, ValueWithContext[AnyRef]]#Context,
+        out: Collector[ValueWithContext[AnyRef]]
+    ): Unit = {
+      val map            = RowConversions.rowToMap(value)
+      val aggregateValue = map.get(aggregateByInternalColumnName).asInstanceOf[AnyRef]
+      val groupValue     = map.get(groupByInternalColumnName)
+      val ctx = pl.touk.nussknacker.engine.api
+        .Context(contextIdGenerator.nextContextId())
+        .withVariable(KeyVariableName, groupValue)
+      val valueWithContext = ValueWithContext(aggregateValue, ctx)
+      out.collect(valueWithContext)
+    }
 
-    // TODO local: calculate based on groupByLazyParam.returnType
-    val groupByFlinkType     = DataTypes.STRING()
-    val aggregateByFlinkType = DataTypes.INT()
-
-    FlinkCustomStreamTransformation((start: DataStream[Context], ctx: FlinkCustomNodeContext) => {
-      val env = start.getExecutionEnvironment
-      // Setting batch mode to enable global window operations. If source is unbounded it will throw a runtime exception
-      env.setRuntimeMode(RuntimeExecutionMode.BATCH)
-
-      val tableEnv = StreamTableEnvironment.create(env)
-
-      val streamOfRows = start.flatMap(new LazyInterpreterFunction(groupByLazyParam, aggregateByLazyParam, ctx))
-
-      val inputParametersTable = NestedRowConversions.buildTableFromRowStream(
-        tableEnv = tableEnv,
-        streamOfRows = streamOfRows,
-        columnSchema = List(
-          ColumnFlinkSchema(groupByInternalColumnName, groupByFlinkType),
-          ColumnFlinkSchema(aggregateByInternalColumnName, aggregateByFlinkType)
-        )
-      )
-
-      val groupedTable = inputParametersTable
-        .groupBy($(groupByInternalColumnName))
-        .select(
-          $(groupByInternalColumnName),
-          call(aggregator.flinkFunctionName, $(aggregateByInternalColumnName)).as(aggregateByInternalColumnName)
-        )
-
-      // TODO local: pass aggregated value and key as separate vars
-      val groupedStream = tableEnv.toDataStream(groupedTable)
-
-      // TODO local: use process function to build a new context with EngineRuntimeContext.contextIdGenerator
-      groupedStream
-        .map(r => {
-          val map = RowConversions.rowToMap(r)
-          ValueWithContext(
-            map.get(aggregateByInternalColumnName).asInstanceOf[AnyRef],
-            Context.withInitialId.withVariable(KeyVariableName, map.get(groupByInternalColumnName))
-          )
-        })
-        .returns(classOf[ValueWithContext[AnyRef]])
-    })
   }
-
-  override def nodeDependencies: List[NodeDependency] = List(OutputVariableNameDependency)
 
   private class LazyInterpreterFunction(
       groupByParam: LazyParameter[AnyRef],
@@ -188,10 +113,10 @@ class TableAggregation
     private lazy val evaluateAggregateByParam = toEvaluateFunctionConverter.toEvaluateFunction(aggregateByParam)
 
     /*
-     Has to out Rows?
-     Otherwise org.apache.flink.util.FlinkRuntimeException: Error during input conversion from external DataStream API to
-     internal Table API data structures. Make sure that the provided data types that configure the converters are
-     correctly declared in the schema.
+         Has to out Rows?
+         Otherwise org.apache.flink.util.FlinkRuntimeException: Error during input conversion from external DataStream API to
+         internal Table API data structures. Make sure that the provided data types that configure the converters are
+         correctly declared in the schema.
      */
     override def flatMap(context: Context, out: Collector[Row]): Unit = {
       collectHandlingErrors(context, out) {
@@ -207,30 +132,4 @@ class TableAggregation
 
   }
 
-}
-
-object Aggregator {
-  val allAggregators: List[Aggregator] = List(Sum, First)
-}
-
-sealed trait Aggregator {
-  val name: String
-  val inputTypeConstraint: Option[TypingResult]
-  def outputType(inputTypeInRuntime: TypingResult): TypingResult
-  val flinkFunctionName: String
-}
-
-case object Sum extends Aggregator {
-  override val name: String = "Sum"
-  // TODO local: is Typed[Number] ok?
-  override val inputTypeConstraint: Option[TypingResult]                  = Some(Typed[Number])
-  override def outputType(inputTypeInRuntime: TypingResult): TypingResult = Typed[Number]
-  override val flinkFunctionName: String                                  = "sum"
-}
-
-case object First extends Aggregator {
-  override val name: String                                               = "First"
-  override val inputTypeConstraint: Option[TypingResult]                  = None
-  override def outputType(inputTypeInRuntime: TypingResult): TypingResult = inputTypeInRuntime
-  override val flinkFunctionName: String                                  = "first_value"
 }
